Clarify text colors
The usage of gfx_SetTextTransparentColor is a little confusing. It
allows you to set a color which will display as transparent when drawing
text. For example, if the transparent color and foreground color are the
same, the text will be transparent. If the transparent color and the
background color are the same, the text will appear with a transparent
background.