More color changes
Changes the background of boxes with letters that are not in the puzzle
word. This makes it much easier to tell when a line has been submitted
when none of the letters are correct. This also aligns with the behavior
of original Wordle.

Also, colors for each state are defined using macros so they can be
changed easily if needed.