Fix error in letter highlighting
Letters were being searched from left to right, so if a "present" letter
appeared before a "correct" letter, the present one would be highlighted
and not the correct one.

This is now fixed by searching for all correct letters first and then
filling in the rest of the present letters.